+++ preload-bugreport.js
@@ -1,5 +1,5 @@
 'use strict';
 const { contextBridge, ipcRenderer } = require('electron');
 contextBridge.exposeInMainWorld('bugAPI', {
-  submit: (payload) => ipcRenderer.invoke('bug-report-submit', payload)
+  openSupport: () => ipcRenderer.send('bug-report-open-support')
 });